# frontend/travel-with-anna-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification for Travel with Anna API",
    "description": "OpenApi documentation for Travel with Anna API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/twa/v1",
      "description": "Local server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/day/generate": {
      "post": {
        "tags": [
          "Day"
        ],
        "operationId": "generateDays",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DayGeneratorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/day/add/{tripId}": {
      "post": {
        "tags": [
          "Day"
        ],
        "operationId": "addDay",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "isFirst",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
...
    "/api/country/cities": {
      "post": {
        "tags": [
          "country-controller"
        ],
        "operationId": "getCountryCities",
        "parameters": [
          {
            "name": "countryName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/City"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/activity": {
      "post": {
        "tags": [
          "Activity"
        ],
        "operationId": "createActivity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/activity/associated": {
      "post": {
        "tags": [
          "Activity"
        ],
        "operationId": "createAssociatedActivities",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityAssociatedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
...
    "/item/save-list": {
      "patch": {
        "tags": [
          "Item"
        ],
        "operationId": "saveAllItemsFromTheList",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/activity/update": {
      "patch": {
        "tags": [
          "Activity"
        ],
        "operationId": "updateActivity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/trip/{tripId}": {
      "get": {
        "tags": [
          "Trip"
        ],
        "operationId": "getTripById",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/item/{backpackId}": {
      "get": {
        "tags": [
          "Item"
        ],
        "operationId": "getAllItemsByBackpackId",
        "parameters": [
          {
            "name": "backpackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/day/{dayId}": {
      "get": {
        "tags": [
          "Day"
        ],
        "operationId": "getDayById",
        "parameters": [
          {
            "name": "dayId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayResponse"
                }
              }
            }
          }
        }
      }
    },
    "/day/trip/{tripId}": {
      "get": {
        "tags": [
          "Day"
        ],
        "operationId": "getDays",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DayResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/activity/trip/{tripId}/details": {
      "get": {
        "tags": [
          "Activity"
        ],
        "operationId": "fetchAddressDetailsByTripId",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressDetail"
                }
              }
            }
          }
        }
      }
    },
    "/activity/day/{dayId}": {
      "get": {
        "tags": [
          "Activity"
        ],
        "operationId": "fetchActivitiesByDayId",
        "parameters": [
          {
            "name": "dayId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityDetailedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/activity/day/{dayId}/details": {
      "get": {
        "tags": [
          "Activity"
        ],
        "operationId": "fetchAddressDetailsByDayId",
        "parameters": [
          {
            "name": "dayId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressDetail"
                }
              }
            }
          }
        }
      }
    },
    "/day/{tripId}": {
      "delete": {
        "tags": [
          "Day"
        ],
        "operationId": "deleteDay",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "isFirst",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/activity/{activityId}": {
      "delete": {
        "tags": [
          "Activity"
        ],
        "operationId": "deleteActivityById",
        "parameters": [
          {
            "name": "activityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ExpanseResponse": {
        "type": "object",
        "properties": {
          "expanseId": {
            "type": "integer",
            "format": "int64"
          },
          "expanseName": {
            "type": "string"
          },
          "expanseCategory": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "paid": {
            "type": "number"
          },
          "exchangeRate": {
            "type": "number"
          },
          "priceInTripCurrency": {
            "type": "number"
          },
          "paidInTripCurrency": {
            "type": "number"
          }
        }
      },
...
      "DayGeneratorRequest": {
        "required": [
          "endDate",
          "startDate"
        ],
        "type": "object",
        "properties": {
          "associatedTripId": {
            "type": "integer",
            "format": "int64"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          }
        }
      },
...
      "City": {
        "type": "object",
        "properties": {
          "city": {
            "type": "string"
          }
        }
      },
      "ActivityRequest": {
        "required": [
          "dateTime",
          "tripId",
          "type"
        ],
        "type": "object",
        "properties": {
          "tripId": {
            "type": "integer",
            "format": "int64"
          },
          "dateTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "activityTitle": {
            "maxLength": 59,
            "minLength": 0,
            "type": "string"
          },
          "badge": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          },
          "type": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          },
          "status": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          },
          "addressRequest": {
            "$ref": "#/components/schemas/AddressRequest"
          },
          "dayTag": {
            "type": "boolean"
          }
        }
      },
      "AddressRequest": {
        "type": "object",
        "properties": {
          "place": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "country": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "countryCode": {
            "maxLength": 3,
            "minLength": 0,
            "type": "string"
          },
          "city": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "phoneNumber": {
            "maxLength": 30,
            "minLength": 0,
            "type": "string"
          },
          "email": {
            "maxLength": 150,
            "minLength": 0,
            "type": "string"
          },
          "currency": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "ActivityAssociatedRequest": {
        "type": "object",
        "properties": {
          "firstRequest": {
            "$ref": "#/components/schemas/ActivityRequest"
          },
          "secondRequest": {
            "$ref": "#/components/schemas/ActivityRequest"
          },
          "addressSeparated": {
            "type": "boolean"
          }
        }
      },
...
      "ItemResponse": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "integer",
            "format": "int64"
          },
          "itemName": {
            "type": "string"
          },
          "qty": {
            "type": "string"
          },
          "isPacked": {
            "type": "boolean"
          },
          "expanse": {
            "$ref": "#/components/schemas/ExpanseResponse"
          }
        }
      },
...
      "ActivityUpdateRequest": {
        "required": [
          "activityId",
          "dayId",
          "newActivityDate",
          "oldActivityDate",
          "startTime"
        ],
        "type": "object",
        "properties": {
          "activityId": {
            "type": "integer",
            "format": "int64"
          },
          "dayId": {
            "type": "integer",
            "format": "int64"
          },
          "oldActivityDate": {
            "type": "string"
          },
          "newActivityDate": {
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "activityTitle": {
            "maxLength": 59,
            "minLength": 0,
            "type": "string"
          },
          "type": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          },
          "addressRequest": {
            "$ref": "#/components/schemas/AddressRequest"
          },
          "dayTag": {
            "type": "boolean"
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "TripResponse": {
        "type": "object",
        "properties": {
          "tripId": {
            "type": "integer",
            "format": "int64"
          },
          "backpackId": {
            "type": "integer",
            "format": "int64"
          },
          "budgetId": {
            "type": "integer",
            "format": "int64"
          },
          "tripName": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "amountOfDays": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "NoteResponse": {
        "type": "object",
        "properties": {
          "noteId": {
            "type": "integer",
            "format": "int64"
          },
          "note": {
            "type": "string"
          }
        }
      },
...
      "ActivityListResponse": {
        "type": "object",
        "properties": {
          "activity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityResponse"
            }
          }
        }
      },
      "ActivityResponse": {
        "type": "object",
        "properties": {
          "activityId": {
            "type": "integer",
            "format": "int64"
          },
          "activityTitle": {
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "badge": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "note": {
            "$ref": "#/components/schemas/NoteResponse"
          },
          "expanse": {
            "$ref": "#/components/schemas/ExpanseResponse"
          },
          "associatedId": {
            "type": "integer",
            "format": "int64"
          },
          "address": {
            "$ref": "#/components/schemas/AddressResponse"
          },
          "dayTag": {
            "type": "boolean"
          }
        }
      },
      "AddressResponse": {
        "type": "object",
        "properties": {
          "addressId": {
            "type": "integer",
            "format": "int64"
          },
          "place": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "countryCode": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          }
        }
      },
      "DayResponse": {
        "type": "object",
        "properties": {
          "dayId": {
            "type": "integer",
            "format": "int64"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "dayOfWeek": {
            "type": "string"
          },
          "isToday": {
            "type": "boolean"
          },
          "dayNumber": {
            "type": "integer",
            "format": "int32"
          },
          "note": {
            "$ref": "#/components/schemas/NoteResponse"
          },
          "tripId": {
            "type": "integer",
            "format": "int64"
          },
          "activities": {
            "$ref": "#/components/schemas/ActivityListResponse"
          }
        }
      },
...
      "Country": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "iso2": {
            "type": "string"
          },
          "iso3": {
            "type": "string"
          }
        }
      },
...
      "AddressDetail": {
        "type": "object",
        "properties": {
          "countries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Country"
            }
          },
          "lastCountry": {
            "$ref": "#/components/schemas/Country"
          },
          "cities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/City"
            }
          },
          "lastCity": {
            "$ref": "#/components/schemas/City"
          }
        }
      },
      "ActivityDetailedResponse": {
        "type": "object",
        "properties": {
          "addressDetail": {
            "$ref": "#/components/schemas/AddressDetail"
          },
          "activities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityResponse"
            }
          },
          "totalPrice": {
            "type": "number"
          },
          "totalPayment": {
            "type": "number"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}